requirements.txt
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
Django==3.2.5
django-cors-headers==3.8.0
djangorestframework==3.12.4
isort==5.9.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
pycodestyle==2.7.0
pylint==2.9.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
